+++ aspecd/tasks.py
@@ -1545,6 +1545,10 @@ class Recipe:
         In case of having a single identifier, use the similar method
         :meth:`aspecd.tasks.Recipe.get_dataset`.
 
+        Datasets are looked up in the :attr:`datasets` and in the
+        :attr:`results` attribute, and (since v0.10) the order of the
+        identifiers preserved in the list of datasets.
+
         Parameters
         ----------
         identifiers : :class:`list`
@@ -1566,13 +1570,17 @@ class Recipe:
         aspecd.tasks.MissingDatasetIdentifierError
             Raised if no identifiers are provided.
 
+
+        .. versionchanged:: 0.10
+            List of datasets preserves list of identifiers.
+
         """
         if not identifiers:
             raise aspecd.exceptions.MissingDatasetIdentifierError
-        matching_datasets = [
-            self.datasets[key] for key in identifiers if key in self.datasets
-        ]
+        matching_datasets = []
         for identifier in identifiers:
+            if identifier in self.datasets:
+                matching_datasets.append(self.datasets[identifier])
             if identifier in self.results:
                 if isinstance(
                     self.results[identifier], aspecd.dataset.Dataset
@@ -1678,6 +1686,10 @@ class Chef:
         aspecd.tasks.MissingRecipeError
             Raised if no recipe is available to be cooked
 
+
+        .. versionchanged:: 0.10
+            All open figures are closed after cooking the recipe.
+
         """
         self._assign_recipe(recipe)
         self._prepare_history()
@@ -1691,6 +1703,7 @@ class Chef:
         self.history["info"]["end"] = datetime.datetime.now().isoformat(
             timespec=self._timespec
         )
+        self._close_figures()
 
     def _assign_recipe(self, recipe):
         if not recipe:
@@ -1721,6 +1734,10 @@ class Chef:
                 else:
                     dataset.replace(source_dir, "")
         self.history["tasks"] = []
+
+    def _close_figures(self):
+        for plotter in self.recipe.plotters.values():
+            plt.close(plotter.figure)
 
 
 class Task(aspecd.utils.ToDictMixin):
@@ -3155,6 +3172,74 @@ class SingleplotTask(PlotTask):
     For more information on the underlying general class,
     see :class:`aspecd.plotting.SinglePlotter`.
 
+
+    Attributes
+    ----------
+    label : :class:`str` or :class:`list`
+        Label for the figure resulting from a plotting step.
+
+        This label will be used to refer to the plot later on when
+        further processing the recipe. Actually, in the recipe's
+        :attr:`Recipe.figures` dict, this label is used as a
+        key and a :obj:`FigureRecord` object stored containing
+        all information necessary for further handling the results of the plot.
+
+        In case of multiple datasets the SingleplotTask is applied to,
+        you can provide as many labels as datasets, to have individual
+        labels for each of the datasets.
+
+        .. note::
+
+            For those being confused what the differences are between
+            :attr:`PlotTask.label` and :attr:`PlotTask.result`: The label
+            refers to the *figure*, *i.e.* the result of a plotting task,
+            whereas the result is a reference to the actual *plotter* that has
+            been used to create the figure referenced with the *label*.
+
+    result : :class:`str` or :class:`list`
+        Label for the plotter of a plotting step.
+
+        This is useful in case of CompositePlotters, where different
+        plotters need to be defined for each of the panels.
+
+        Internally, this label is used as key in the recipe's
+        :attr:`Recipe.plotters` dict to store the actual
+        :class:`aspecd.plotting.Plotter` object.
+
+        In case of multiple datasets the SingleplotTask is applied to,
+        you can provide as many result labels as datasets, to have individual
+        labels referring to the individual plotter for each of the datasets.
+
+        .. note::
+
+            For those being confused what the differences are between
+            :attr:`PlotTask.label` and :attr:`PlotTask.result`: The label
+            refers to the *figure*, *i.e.* the result of a plotting task,
+            whereas the result is a reference to the actual *plotter* that has
+            been used to create the figure referenced with the *label*.
+
+    target : :class:`str`
+        Label of an existing previous plotter the plot should be added to.
+
+        Sometimes it is desirable to add something to an already existing
+        plot after this original plot has been created. Programmatically,
+        this would be equivalent to setting the
+        :attr:`aspecd.plotting.Plotter.figure` and
+        :attr:`aspecd.plotting.Plotter.axes` attributes of the underlying
+        plotter object.
+
+        The result: Your plot will be a new figure window, but with the
+        original plot contained and the new plot added on top of it.
+
+    annotations : :class:`list`
+        Labels of plot annotations that should be applied to the plot.
+
+        The labels need to be valid keys of the :attr:`Recipe.annotations`
+        attribute.
+
+
+    Examples
+    --------
     For an example of how such a singleplot task may be included into a recipe,
     see the YAML listing below:
 
@@ -3236,21 +3321,6 @@ class SingleplotTask(PlotTask):
     ``settings`` dict of your recipe and set it to False.
 
     """
-
-    def perform(self):
-        """
-        Call the appropriate method of the underlying object.
-
-        For details, see the method :meth:`aspecd.tasks.Task.perform` of the
-        base class.
-
-        Additionally, to what is done in the base class, a PlotTask adds a
-        :obj:`aspecd.tasks.FigureRecord` object to the
-        :attr:`aspecd.tasks.Recipe.figures` property of the underlying
-        recipe in case an :attr:`aspecd.tasks.PlotTask.label` has been set.
-
-        """
-        super().perform()
 
     def _add_figure_to_recipe(self, label=""):
         figure_record = FigureRecord()